requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-bootstrap5==21.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-filter==21.1
django-import-export==2.6.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.8
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
